build: coverage for integration tests
