+++ examples/spock/src/main/java/io/github/joke/examples/Person.java
@@ -1,4 +1,4 @@
-package io.github.joke.spockmockable.tests;
+package io.github.joke.examples;
 
 public final class Person {
