+++ examples/spock/src/main/java/io/github/joke/examples/StaticCaller.java
@@ -1,4 +1,4 @@
-package io.github.joke.spockmockable.tests;
+package io.github.joke.examples;
 
 public class StaticCaller {
