+++ examples/spock/src/test/groovy/io/github/joke/examples/StaticTest.groovy
@@ -1,4 +1,4 @@
-package io.github.joke.spockmockable.tests
+package io.github.joke.examples
 
 import spock.lang.Specification
